+++ darwin/photo_manager/Sources/photo_manager/core/PMBaseFilter.h
@@ -13,4 +13,6 @@
 
 - (BOOL) needTitle;
 
+- (BOOL) onlyLocal;
+
 @end