+++ darwin/photo_manager/Sources/photo_manager/core/PMConvertUtils.m
@@ -144,6 +144,7 @@
         container.containsLivePhotos = [map[@"containsLivePhotos"] boolValue];
         container.onlyLivePhotos = [map[@"onlyLivePhotos"] boolValue];
         container.includeHiddenAssets = [map[@"includeHiddenAssets"] boolValue];
+        container.onlyLocal = [map[@"onlyLocal"] boolValue];
         
         NSArray *sortArray = map[@"orders"];
         [container injectSortArray:sortArray];