+++ darwin/photo_manager/Sources/photo_manager/core/PMFilterOption.h
@@ -59,6 +59,7 @@ typedef struct PMDurationConstraint {
 @property(nonatomic, assign) BOOL onlyLivePhotos;
 @property(nonatomic, assign) BOOL containsModified;
 @property(nonatomic, assign) BOOL includeHiddenAssets;
+@property(nonatomic, assign) BOOL onlyLocal;
 @property(nonatomic, strong) NSArray<NSSortDescriptor *> *sortArray;
 
 - (NSArray<NSSortDescriptor *> *)sortCond;